requirements.txt
Django==1.8.2
mock==1.0.1
Pillow==3.0.0
py==1.4.28
pytest==2.7.1
pytz==2015.7
wheel==0.24.0
